+++ src/lua/CommunityBalanceMod/FortressPvE/Balance.lua
@@ -3,8 +3,8 @@
 -- FortressPvE
 kFortressUpgradeCost = 24
 kFortressResearchTime = 25
-kFortressAbilityCooldown = 10
-kFortressAbilityCost = 3
+kFortressAbilityCooldown = 15
+kFortressAbilityCost = 2
 kCragCost = 8
 kShiftCost = 8
 kShadeCost = 8
@@ -14,12 +14,14 @@ kWhipCost = 8
 kCragUmbra = 5
 
 kMaxHallucinations = 6
-kHallucinationLifeTime = 600
+kHallucinationLifeTime = 0.1 -- ignored, last indefinitely
 
 kStormCloudDuration = 9.5
 
-kWhipAbilityCost = 3
+kWhipAbilityCost = 2
 kWhipAbilityCooldown = 10
 
 kHallucinateCloningCost = 0
-kHallucinateCloningCooldown = 1
+kHallucinateCloningCooldown = 1.5
+kHallucinateRandomCost = 0
+kHallucinateRandomCooldown = 1.5