+++ src/lua/CommunityBalanceMod/FortressPvE/Shift.lua
@@ -21,6 +21,17 @@ function Shift:OnCreate()
 end
 
 
+function Shift:GetOffInfestationHurtPercentPerSecond()
+
+    if self:GetTechId() == kTechId.FortressShift then 
+        return kBalanceOffInfestationHurtPercentPerSecondFortress
+    end
+
+    return kBalanceOffInfestationHurtPercentPerSecond
+end
+
+
+
 function Shift:GetMaxSpeed()
 
     if self:GetTechId() == kTechId.FortressShift then
@@ -279,7 +290,7 @@ function Shift:GetTechAllowed(techId, techNode, player)
     
             allowed = allowed and not GetHasTech(self, kTechId.FortressShift) and not  GetIsTechResearching(self, techId)
         elseif techId == kTechId.FortressShiftAbility then
-            allowed = self:GetTechId() == kTechId.FortressShift and GetHasTech(self, kTechId.ShiftHive)
+            allowed = allowed and ( self:GetTechId() == kTechId.FortressShift ) and GetHasTech(self, kTechId.ShiftHive)
         else